requirements.txt
streamlit==1.31.1
streamlit-image-comparison==0.0.4
pandas==2.2.0
numpy==1.26.4
pillow==10.2.0
torch==2.2.0
opencv-python==4.9.0.80